+++ packages/client/src/views/settings/hotkeys/index.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 
 import { Hotkey } from '$components/hotkey';
 import { TabPanel } from '$components/tabs';
-import { VarHotkey } from '$components/var/VarHotkey';
+import { VarCategory, VarHotkey, VarScope } from '$components/var';
 import { hotkeyGroups } from '$data/hotkeys';
 import styles from './index.module.scss';
 
@@ -11,22 +11,27 @@ export const SettingsHotkeys: React.FC = observer(() => {
   return (
     <TabPanel id="hotkeys">
       <h2>Keyboard shortcuts</h2>
-      {Object.entries(hotkeyGroups).map(([groupId, group]) => (
-        <div key={groupId} className={styles.group}>
-          <h3>{group.label}</h3>
-          {Object.entries(group.hotkeys).map(([hotkeyId, hotkey]) => (
-            <div key={hotkeyId} id={hotkeyId} className={styles.hotkey}>
-              <div>{hotkey.label}</div>
-              <div>
-                <Hotkey keys={hotkey.defaultKeys} />
+      <VarScope path="app.hotkeys">
+        {Object.entries(hotkeyGroups).map(([groupId, group]) => (
+          <VarCategory
+            label={group.label}
+            key={groupId}
+            className={styles.group}
+          >
+            {Object.entries(group.hotkeys).map(([hotkeyId, hotkey]) => (
+              <div key={hotkeyId} id={hotkeyId} className={styles.hotkey}>
+                <div>{hotkey.label}</div>
+                <div>
+                  <Hotkey keys={hotkey.defaultKeys} />
+                </div>
+                <div>
+                  <VarHotkey path={`${groupId}_${hotkeyId}`} />
+                </div>
               </div>
-              <div>
-                <VarHotkey path={`app.hotkeys.${groupId}_${hotkeyId}`} />
-              </div>
-            </div>
-          ))}
-        </div>
-      ))}
+            ))}
+          </VarCategory>
+        ))}
+      </VarScope>
     </TabPanel>
   );
 });